Feature/117 add logging(visitor controller,car controller,crossing controller)

+++ src/main/java/ru/ac/checkpointmanager/controller/VisitorController.java
@@ -21,6 +21,7 @@ import ru.ac.checkpointmanager.utils.ErrorUtils;
 import ru.ac.checkpointmanager.utils.Mapper;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @RestController
@@ -43,12 +44,13 @@ public class VisitorController {
     })
     @PostMapping
     public ResponseEntity<?> addVisitor(@Valid @RequestBody VisitorDTO visitorDTO, BindingResult bindingResult) {
-
         if (bindingResult.hasErrors()) {
+            log.warn("Failed to add visitor due to validation errors");
             return new ResponseEntity<>(ErrorUtils.errorsList(bindingResult), HttpStatus.BAD_REQUEST);
         }
 
         Visitor newVisitor = visitorService.addVisitor(mapper.toVisitor(visitorDTO));
+        log.info("New visitor added: {}", newVisitor);
         return new ResponseEntity<>(mapper.toVisitorDTO(newVisitor), HttpStatus.CREATED);
     }
 
@@ -59,8 +61,10 @@ public class VisitorController {
     @GetMapping("/{id}")
     public ResponseEntity<?> getVisitor(@PathVariable UUID id) {
         Visitor existVisitor = visitorService.getVisitor(id);
+        log.debug("Retrieved visitor with ID {}", id);
         return new ResponseEntity<>(mapper.toVisitorDTO(existVisitor), HttpStatus.OK);
     }
+
 
     @Operation(summary = "Обновить информацию о посетителе по ID")
     @ApiResponses(value = {
@@ -69,15 +73,18 @@ public class VisitorController {
     })
     @PutMapping("/{id}")
     public ResponseEntity<?> updateVisitor(@PathVariable UUID id, @RequestBody VisitorDTO visitorDTO,
-                                          BindingResult bindingResult) {
+                                           BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
+            log.warn("Failed to update visitor due to validation errors");
             return new ResponseEntity<>(ErrorUtils.errorsList(bindingResult), HttpStatus.BAD_REQUEST);
         }
 
         Visitor visitor = mapper.toVisitor(visitorDTO);
-        Visitor updateVisitor = visitorService.updateVisitor(id, visitor);
-        return new ResponseEntity<>(mapper.toVisitorDTO(updateVisitor), HttpStatus.OK);
+        Visitor updatedVisitor = visitorService.updateVisitor(id, visitor);
+        log.info("Visitor updated with ID {}", id);
+        return new ResponseEntity<>(mapper.toVisitorDTO(updatedVisitor), HttpStatus.OK);
     }
+
 
     @Operation(summary = "Удалить посетителя по ID")
     @ApiResponses(value = {
@@ -85,7 +92,12 @@ public class VisitorController {
     })
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteVisitor(@PathVariable UUID id) {
+        if (!visitorService.existsById(id)) {
+            log.warn("Failed to delete visitor: No visitor found with ID {}", id);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         visitorService.deleteVisitor(id);
+        log.info("Visitor with ID {} deleted", id);
         return new ResponseEntity<>("Visitor deleted " + id, HttpStatus.OK);
     }
 
@@ -94,11 +106,12 @@ public class VisitorController {
             @ApiResponse(responseCode = "200", description = "Посетители найдены"),
     })
     @GetMapping("/phone")
-    public ResponseEntity<List<VisitorDTO>> searchByPhone(@RequestParam String phone) {
+    public List<VisitorDTO> searchByPhone(@RequestParam String phone) {
         List<Visitor> visitors = visitorService.findByPhonePart(phone);
-        List<VisitorDTO> visitorDTOS = mapper.toVisitorDTOS(visitors);
-        return new ResponseEntity<>(visitorDTOS, HttpStatus.OK);
+        log.debug("Visitors found with phone part: {}", phone);
+        return mapper.toVisitorDTOS(visitors);
     }
+
 
     @Operation(summary = "Найти посетителя по имени")
     @ApiResponses(value = {
@@ -107,9 +120,10 @@ public class VisitorController {
     @GetMapping("/name")
     public ResponseEntity<List<VisitorDTO>> searchByName(@RequestParam String name) {
         List<Visitor> visitors = visitorService.findByNamePart(name);
-        List<VisitorDTO> visitorDTOS = mapper.toVisitorDTOS(visitors);
-        return new ResponseEntity<>(visitorDTOS, HttpStatus.OK);
+        log.debug("Visitors found with name part: {}", name);
+        return new ResponseEntity<>(mapper.toVisitorDTOS(visitors), HttpStatus.OK);
     }
+
 
     @Operation(summary = "Найти посетителя по ID пропуска")
     @ApiResponses(value = {
@@ -119,11 +133,10 @@ public class VisitorController {
     @GetMapping("/pass")
     public ResponseEntity<?> searchByPass(@RequestParam UUID uuid) {
         Visitor existVisitor = visitorService.findByPassId(uuid).orElse(null);
-        if (uuid != null) {
-            return new ResponseEntity<>(mapper.toVisitorDTO(existVisitor), HttpStatus.OK);
-        }
-        return new ResponseEntity<>("There is no such visitor in any pass! ", HttpStatus.NOT_FOUND);
+        log.debug("Visitor found for pass ID: {}", uuid);
+        return new ResponseEntity<>(mapper.toVisitorDTO(existVisitor), HttpStatus.OK);
     }
+
 
     @Operation(summary = "Найти посетителей из пропусков пользователя")
     @ApiResponses(value = {
@@ -133,7 +146,8 @@ public class VisitorController {
     @GetMapping("/user/{userId}")
     public ResponseEntity<List<VisitorDTO>> searchByUserId(@PathVariable UUID userId) {
         List<Visitor> visitors = visitorService.findByUserId(userId);
-        List<VisitorDTO> visitorDTOS = mapper.toVisitorDTOS(visitors);
-        return new ResponseEntity<>(visitorDTOS, HttpStatus.OK);
+        log.debug("Visitors found for user ID: {}", userId);
+        return new ResponseEntity<>(mapper.toVisitorDTOS(visitors), HttpStatus.OK);
     }
+
 }
